Restructure linfa-linear crate
