+++ linfa-linear/src/glm.rs
@@ -1,10 +1,1 @@
 mod link;
-
-use ndarray::NdFloat;
-use ndarray_linalg::Lapack;
-use num_traits::FromPrimitive;
-
-pub trait Float: NdFloat + Lapack + Default + Clone + FromPrimitive {}
-
-impl Float for f32 {}
-impl Float for f64 {}
